cloud_run_functions/requirements.txt
functions-framework==3.*
pandas==2.2.2
pandas_gbq==0.23.1
google-auth==2.34.0
gspread==6.2.1
serpapi==0.1.5
google-cloud-bigquery==3.25.0
google-search-results==2.4.2
requests==2.32.3
deep-translator==1.11.4
